backend/requirements.txt
# FILE: requirements.txt

Flask
Flask-Cors
mysql-connector-python
pytz
watchdog
